Release version 0.5.0

+++ lib/controller/conversation/zap_share_controller.dart
@@ -1,19 +1,16 @@
 import 'dart:async';
 import 'dart:convert';
-import 'dart:io';
 import 'dart:typed_data';
 
-import 'package:archive/archive_io.dart';
 import 'package:chat_interface/connection/connection.dart';
 import 'package:chat_interface/connection/encryption/symmetric_sodium.dart';
 import 'package:chat_interface/connection/messaging.dart';
-import 'package:chat_interface/controller/conversation/attachment_controller.dart';
 import 'package:chat_interface/controller/conversation/conversation_controller.dart';
-import 'package:chat_interface/controller/conversation/message_controller.dart' as msg;
+import 'package:chat_interface/controller/conversation/message_controller.dart';
+import 'package:chat_interface/controller/conversation/message_provider.dart' as msg;
 import 'package:chat_interface/controller/current/status_controller.dart';
 import 'package:chat_interface/main.dart';
 import 'package:chat_interface/pages/chat/components/conversations/zap_share_window.dart';
-import 'package:chat_interface/pages/chat/components/message/message_feed.dart';
 import 'package:chat_interface/theme/ui/dialogs/window_base.dart';
 import 'package:chat_interface/util/logging_framework.dart';
 import 'package:chat_interface/util/popups.dart';
@@ -23,6 +20,7 @@ import 'package:file_selector/file_selector.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:get/get.dart';
+import 'package:liphium_bridge/liphium_bridge.dart';
 import 'package:open_app_file/open_app_file.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:sodium_libs/sodium_libs.dart';
@@ -42,9 +40,8 @@ class ZapShareController extends GetxController {
   String? transactionToken;
   String? uploadToken;
   String? filePath;
-  Directory? chunksDir;
+  XDirectory? chunksDir;
   SecureKey? key;
-  Timer? zapper;
   StreamSubscription<Uint8List>? partSubscription;
 
   static const chunkSize = 512 * 1024;
@@ -65,6 +62,7 @@ class ZapShareController extends GetxController {
     transactionToken = null;
     uploadToken = null;
     filePath = null;
+    zapperStarted = false;
 
     // Cancel the partSubscription if it exists and set it to null
     partSubscription?.cancel();
@@ -77,7 +75,7 @@ class ZapShareController extends GetxController {
     }
     if (uploading) {
       connector.sendAction(
-        Message("cancel_transaction", <String, dynamic>{}),
+        ServerAction("cancel_transaction", <String, dynamic>{}),
       );
     } else {
       partSubscription?.cancel();
@@ -87,25 +85,29 @@ class ZapShareController extends GetxController {
 
   /// Open the window for zap share for a conversation
   void openWindow(Conversation conversation, ContextMenuData data) async {
-    // If Zap Share is already doing something, show a menu
+    // If Zap is already doing something, show a menu
     if (isRunning()) {
       Get.dialog(ZapShareWindow(data: data, conversation: conversation));
       return;
     }
 
-    // Grab files
-    final files = await openFiles();
-    if (files.isEmpty) {
+    // If Zap is not running, let them select a file for transfer
+    final file = await openFile();
+    if (file == null) {
       return;
     }
 
     // Start the transaction
-    newTransaction(conversation.otherMember.id, conversation.id, files);
+    newTransaction(conversation.otherMember.id, conversation.id, [file]);
   }
 
   //* Everything about sending starts here
 
   void newTransaction(LPHAddress friend, LPHAddress conversationId, List<XFile> files) async {
+    if (files.length > 1) {
+      sendLog("zapping multiple files is currently not supported");
+      return;
+    }
     if (isRunning()) {
       sendLog("Already in a transaction");
       return;
@@ -117,35 +119,16 @@ class ZapShareController extends GetxController {
     sending = 0;
 
     // Generate chunks dir
-    var tempDir = await getTemporaryDirectory();
-    tempDir = Directory(path.join(tempDir.path, "liphium"));
+    var tempPath = await getTemporaryDirectory();
+    final tempDir = XDirectory(path.join(tempPath.path, "liphium"));
     await tempDir.create();
     chunksDir = await tempDir.createTemp("zapzap");
     await chunksDir!.create();
     sendLog(chunksDir!.path);
 
-    // If there are more files than one, put them into an archive
+    // If there are more files than one, put them into an archive (TODO: Implement)
     step.value = "chat.zapshare.compressing".tr;
-    File file;
-    if (files.length > 1) {
-      // TODO: Have this in an isolate, when https://github.com/brendan-duncan/archive/issues/349 gets an answer (or turn this into rust code)
-
-      // Start the encoder
-      final outFile = path.join(chunksDir!.path, "zapped_files.zip");
-      final encoder = ZipFileEncoder();
-      encoder.createWithBuffer(OutputFileStream(outFile));
-
-      // Add all files to the archive
-      for (var file in files) {
-        await encoder.addFile(File(file.path));
-      }
-
-      // Run all the compression in an isolate
-      await encoder.close();
-      file = File(outFile);
-    } else {
-      file = File(files[0].path);
-    }
+    XFile file = files[0];
 
     // Intialize the encryption key
     key = randomSymmetricKey();
@@ -157,7 +140,7 @@ class ZapShareController extends GetxController {
     endPart = (fileSize.toDouble() / chunkSize.toDouble()).ceil();
     step.value = "chat.zapshare.waiting".tr;
     connector.sendAction(
-      Message("create_transaction", <String, dynamic>{
+      ServerAction("create_transaction", <String, dynamic>{
         "name": fileName,
         "size": fileSize,
       }),
@@ -176,7 +159,7 @@ class ZapShareController extends GetxController {
 
         // Send live share message
         final container = LiveshareInviteContainer(event.data["url"], transactionId!, transactionToken!, fileName, key!);
-        sendActualMessage(false.obs, conversationId, msg.MessageType.liveshare, [], container.toJson(), "", () => {});
+        Get.find<MessageController>().currentProvider.value!.sendMessage(false.obs, msg.MessageType.liveshare, [], container.toJson(), "");
       },
     );
   }
@@ -184,12 +167,14 @@ class ZapShareController extends GetxController {
   // For the zapper to know what to download
   int currentlySending = 0;
   int currentEndPart = 0;
+  bool zapperStarted = false;
 
   /// Zap deamon
-  void _startZapper(int start, int end) {
-    if (zapper != null) {
+  void _startZapper(int start, int end) async {
+    if (zapperStarted) {
       return;
     }
+    zapperStarted = true;
     waiting.value = false;
     step.value = "chat.zapshare.uploading".tr;
 
@@ -199,23 +184,23 @@ class ZapShareController extends GetxController {
     bool uploaded = true;
     int tries = 0;
 
-    // Start a new periodic timer (no while true loop cause it's not needed, a fast timer will be enough)
-    zapper = Timer.periodic(10.ms, (timer) async {
+    while (true) {
       // Cancel if the thing isn't running anymore
       if (!isRunning()) {
-        timer.cancel();
+        break;
       }
 
       // Cancel if it took to many tries to upload the current part
       if (tries > 5) {
-        timer.cancel();
+        showErrorPopup("zap.error", "server.error".tr);
         cancel();
-        return;
+        break;
       }
 
       // Check if new parts can be sent
       if (currentlySending >= currentEndPart || !uploaded) {
-        return;
+        await Future.delayed(10.ms); // To prevent infinite spinning
+        continue;
       }
 
       // Update all the variables to send the new part
@@ -229,48 +214,48 @@ class ZapShareController extends GetxController {
       if (!success) {
         // Retry in case of an error and wait a little bit
         currentlySending--;
-        showErrorPopup("error", "server.error".tr);
         tries++;
         await Future.delayed(2000.ms);
       } else {
         tries = 0;
       }
       uploaded = true;
-    });
+    }
   }
 
   /// Upload the actual file part to the server
   Future<bool> _sendActualFilePart(int chunk) async {
     if (!isRunning()) {
-      sendLog("why would the server ask for parts when zap share isn't even running :smug:");
+      sendLog("why would the server ask for parts when zap isn't even running :smug:");
       return false;
     }
 
     // Read the original file and extract one chunk
-    final file = File(filePath!);
+    final file = XFile(filePath!);
     final stream = file.openRead((chunk - 1) * chunkSize, chunk * chunkSize);
-    final chunkFile = File("${chunksDir!.path}/chunk_$chunk");
-    await chunkFile.create();
-    final toEncrypt = <int>[];
+    var currentIndex = 0;
+    final toEncrypt = Uint8List(chunkSize);
 
     // Send the chunk once done
     final completer = Completer<bool>();
     stream.listen(
       (bytes) {
-        toEncrypt.addAll(bytes);
+        for (var byte in bytes) {
+          toEncrypt[currentIndex] = byte;
+          currentIndex++;
+        }
       },
       cancelOnError: true,
       onError: (e) => sendLog(e),
       onDone: () async {
         // Encrypt all the bytes and write them to the chunk file
         final encrypted = encryptSymmetricBytes(Uint8List.fromList(toEncrypt), key!);
-        await chunkFile.writeAsBytes(encrypted.toList());
 
         // Send chunk
         final formData = d.FormData.fromMap({
           "id": transactionId,
           "token": uploadToken,
-          "part": await d.MultipartFile.fromFile(chunkFile.path),
+          "part": d.MultipartFile.fromBytes(encrypted, filename: "chunk_$chunk"),
         });
         final res = await dio.post(
           nodePath("/auth/liveshare/upload"),
@@ -290,7 +275,6 @@ class ZapShareController extends GetxController {
         }
 
         if (res.statusCode != 200) {
-          showErrorPopup("error", "server.error".tr);
           sendLog("Failed to send chunk $chunk");
           completer.complete(false);
           return;
@@ -298,7 +282,7 @@ class ZapShareController extends GetxController {
 
         final json = jsonDecode(res.data);
         if (!json["success"]) {
-          showErrorPopup("error", json["error"]);
+          sendLog("Failed to send chunk $chunk cause ${json["error"]}");
           completer.complete(false);
           return;
         }
@@ -341,18 +325,19 @@ class ZapShareController extends GetxController {
     transactionId = null;
     transactionToken = null;
     filePath = null;
-    zapper?.cancel();
-    zapper = null;
+    zapperStarted = false;
     uploadToken = null;
-    Timer.periodic(2000.ms, (timer) async {
-      try {
-        await chunksDir?.delete(recursive: true);
-        chunksDir = null;
-        timer.cancel();
-      } catch (e) {
-        sendLog("couldn't delete chunk directory: $e");
-      }
-    });
+    if (chunksDir != null) {
+      Timer.periodic(2000.ms, (timer) async {
+        try {
+          await chunksDir?.delete(recursive: true);
+          chunksDir = null;
+          timer.cancel();
+        } catch (e) {
+          sendLog("couldn't delete chunk directory: $e");
+        }
+      });
+    }
     resetControllerState();
   }
 
@@ -383,7 +368,15 @@ class ZapShareController extends GetxController {
     }
     endPart = (json["size"].toDouble() / chunkSize.toDouble()).ceil();
 
-    sendLog(base64Encode(container.key.extractBytes()));
+    // Let the user decide where to save the file
+    // TODO: Think about something else for this on mobile
+    final FileSaveLocation? receiveFile = await getSaveLocation(suggestedName: container.fileName);
+    if (receiveFile == null) {
+      showErrorPopup("error", "zap.no_save_location".tr);
+      return;
+    }
+
+    // Set state visible in the app
     key = container.key;
     currentConversation.value = conversation;
     currentReceiver.value = friendAddress;
@@ -406,24 +399,16 @@ class ZapShareController extends GetxController {
     );
     final body = res.data as d.ResponseBody;
 
-    // Create receiving directory
-    var tempDir = await getTemporaryDirectory();
-    tempDir = Directory(path.join(tempDir.path, "liphium"));
-    await tempDir.create();
-    final receiveDir = await tempDir.createTemp("liveshare-recv");
-    await receiveDir.create();
-    sendLog(receiveDir.path);
-
     // Data for downloads
-    bool completed = false, downloading = false, waiting = false;
+    bool completed = false, waiting = false;
     int currentChunk = 0;
     int maxChunk = 0;
+    int tries = 0;
     String receiverId = "";
     bool receivedInfo = false;
 
     // Listen for new parts
     step.value = "chat.zapshare.downloading".tr;
-    Timer? downloadTimer;
     partSubscription = body.stream.listen(
       (event) async {
         // Parse the server sent event (Format: data: <data>\n\n)
@@ -439,12 +424,16 @@ class ZapShareController extends GetxController {
         if (currentChunk == 0) {
           currentChunk = maxChunk;
 
-          // Start a timer for downloading the parts
-          downloadTimer = Timer.periodic(10.ms, (timer) async {
-            if (downloading) return;
+          while (true) {
             if (completed) {
-              timer.cancel();
-              return;
+              break;
+            }
+
+            // Only try a max of 5 times
+            if (tries > 5) {
+              showErrorPopup("zap.error", "server.error".tr);
+              cancel();
+              break;
             }
 
             // Check if new chunk is available
@@ -454,10 +443,10 @@ class ZapShareController extends GetxController {
                 progress.value = currentChunk / endPart;
                 waiting = false;
               } else {
-                return;
+                await Future.delayed(10.ms); // To prevent infinite spinning
+                continue;
               }
             }
-            downloading = true;
 
             // Download stuff
             final formData = d.FormData.fromMap({
@@ -465,23 +454,27 @@ class ZapShareController extends GetxController {
               "token": container.token,
               "chunk": currentChunk,
             });
-            final res = await dio.download(
+            final res = await dio.get<Uint8List>(
               "${nodeProtocol()}${container.url}/liveshare/download",
-              "${receiveDir.path}/chunk_$currentChunk",
               data: formData,
               options: d.Options(
+                responseType: d.ResponseType.bytes,
                 validateStatus: (status) => true,
               ),
             );
 
             if (res.statusCode != 200) {
               sendLog("ERROR DOWNLOADING CHUNK $currentChunk");
-              showErrorPopup("error", "server.error".tr);
-              return;
+              await Future.delayed(2000.ms); // Wait a little before trying again
+              tries++;
+              continue;
             }
-
-            downloading = false;
+            tries = 0;
             currentPart.value = currentChunk;
+
+            // Append the chunk to the file
+            final decrypted = decryptSymmetricBytes(res.data!, key!);
+            await fileUtil.appendToFile(XFile(receiveFile.path), decrypted);
 
             // Check if new chunk can be downloaded right away
             if (currentChunk < maxChunk) {
@@ -491,6 +484,7 @@ class ZapShareController extends GetxController {
               waiting = true;
             }
 
+            // Tell the server the part was successfully received
             _tellReceived(
               container.url,
               container.id,
@@ -499,18 +493,14 @@ class ZapShareController extends GetxController {
               callback: (complete) async {
                 completed = complete;
                 if (completed) {
-                  sendLog("download completed, stitching..");
-                  final dir = await _stitchFileTogether(container.fileName, receiveDir);
-                  await receiveDir.delete(recursive: true);
-                  if (dir != null) {
-                    OpenAppFile.open(dir);
-                  }
+                  sendLog("download with zap completed, opening final folder..");
+                  OpenAppFile.open(path.dirname(receiveFile.path));
                   partSubscription?.cancel();
                 }
               },
               onError: () => sendLog("error"),
             );
-          });
+          }
         }
       },
       onError: (e) {
@@ -518,14 +508,7 @@ class ZapShareController extends GetxController {
       },
       cancelOnError: true,
       onDone: () async {
-        downloadTimer?.cancel();
         onTransactionEnd();
-        // Give time to delete files
-        Timer(const Duration(seconds: 5), () {
-          if (!completed) {
-            receiveDir.delete(recursive: true);
-          }
-        });
       },
     );
   }
@@ -555,50 +538,6 @@ class ZapShareController extends GetxController {
     }
 
     callback?.call(res.data["complete"]);
-  }
-
-  /// Returns the directory the file was saved to (if successful)
-  Future<String?> _stitchFileTogether(String fileName, Directory dir) async {
-    step.value = "chat.zapshare.finishing".tr;
-
-    // Let the user pick where to store the finished file on desktop
-    File file;
-    if (GetPlatform.isDesktop) {
-      final FileSaveLocation? result = await getSaveLocation(suggestedName: fileName);
-      if (result == null) {
-        showErrorPopup("error", "zap.no_save_location".tr);
-        file = File(path.join(AttachmentController.getFilePathForType(StorageType.permanent), fileName));
-      } else {
-        file = File(result.path);
-      }
-    } else {
-      // Just put it into the downloads folder
-      // TODO: Check if this actually works on mobile
-      final downloads = await getDownloadsDirectory();
-      file = File(path.join(downloads!.path, fileName));
-    }
-
-    // Stitch together the final file
-    int currentIndex = 1;
-    while (true) {
-      final chunk = File("${dir.path}/chunk_$currentIndex");
-      final exists = await chunk.exists();
-      if (!exists) {
-        break;
-      }
-
-      // Decrypt the current chunk
-      final bytes = await chunk.readAsBytes();
-      final decrypted = decryptSymmetricBytes(bytes, key!);
-
-      // Add the chunk to the file and delete the chunk
-      await file.writeAsBytes(decrypted, mode: currentIndex == 1 ? FileMode.write : FileMode.append);
-      chunk.delete();
-
-      currentIndex++;
-    }
-
-    return file.parent.path;
   }
 }
