+++ lib/pages/status/setup/settings_setup.dart
@@ -1,4 +1,5 @@
-import 'package:chat_interface/pages/settings/app/language_settings.dart';
+import 'package:chat_interface/main.dart';
+import 'package:chat_interface/pages/settings/app/general_settings.dart';
 import 'package:chat_interface/pages/settings/app/log_settings.dart';
 import 'package:chat_interface/pages/settings/appearance/theme_settings.dart';
 import 'package:chat_interface/pages/settings/data/settings_controller.dart';
@@ -22,7 +23,7 @@ class SettingsSetup extends Setup {
     }
 
     // Set current language
-    Get.updateLocale(LanguageSettings.languages[controller.settings[LanguageSettings.language]!.getValue()].locale);
+    Get.updateLocale(GeneralSettings.languages[controller.settings[GeneralSettings.language]!.getValue()].locale);
 
     // Changes the color theme
     Get.find<ThemeManager>().changeTheme(getThemeData());
@@ -31,14 +32,16 @@ class SettingsSetup extends Setup {
     TabletopSettings.initSettings();
 
     // Delete old logs
-    final list = await LogManager.loggingDirectory!.list().toList();
-    list.sort((a, b) => a.statSync().modified.compareTo(b.statSync().modified));
-    var index = Get.find<SettingController>().settings[LogSettings.amountOfLogs]!.getValue() as double;
-    for (final file in list) {
-      if (index <= 0) {
-        await file.delete();
+    if (!isWeb) {
+      final list = await LogManager.loggingDirectory!.list().toList();
+      list.sort((a, b) => a.statSync().modified.compareTo(b.statSync().modified));
+      var index = Get.find<SettingController>().settings[LogSettings.amountOfLogs]!.getValue() as double;
+      for (final file in list) {
+        if (index <= 0) {
+          await file.delete();
+        }
+        index--;
       }
-      index--;
     }
 
     sendLog("hi hi");
